refactored searchboxwithfilter—decoupled filter component to prepare for reuse
