+++ app/client/components/SearchBoxWithFilter.jsx
@@ -1,24 +1,21 @@
 import React from "react";
 import { connect } from 'react-redux';
-import { ModuleBox, StyledSelect, Input, StyledButton} from "./styledElements";
+import { ModuleBox, Input} from "./styledElements";
 import styled from "styled-components";
 import matchSorter from "match-sorter";
 import PropTypes from 'prop-types';
 
-import { replaceUnderscores } from "../utils";
 import {loadFilters, createFilter} from "../reducers/filterReducer";
 import LoadingSpinner from "./loadingSpinner.jsx";
 
+import Filter from './filter';
+
 const propTypes = {
-  onFiltersChanged: PropTypes.func,
   withSearch: PropTypes.bool,
   toggleTitle: PropTypes.string
 }
 
 const defaultProps = {
-  onFiltersChanged: (filter) => {
-    console.log("Filter applied:", filter)
-  },
   withSearch: true,
   toggleTitle: "Filters"
 }
@@ -43,29 +40,12 @@ class SearchBoxWithFilter extends React.Component {
     this.state = {
       phrase: '',
       filters: [],
-      keyName: 'select',
-      value: 'select',
-      filterType: 'of',
-      keyNameValues: {},
       possibleFilters: null,
       filterActive: false,
       dropdownActive: false,
       saveFilterDropdown: false,
-      loadFiltersDropdown: false,
-      filterName: null,
-      attributeListIsFullyFloatParsable: false,
-      GtLtThreshold: 0
+      loadFiltersDropdown: false
     };
-    this.keyNameChange = this.keyNameChange.bind(this);
-    this.handleChange = this.handleChange.bind(this);
-    this.addFilter = this.addFilter.bind(this);
-    this.removeFilter = this.removeFilter.bind(this);
-    this.toggleFilter = this.toggleFilter.bind(this);
-    this.dropdownActive = this.dropdownActive.bind(this);
-    this.saveFilter = this.saveFilter.bind(this);
-    this.saveFilterDropdown = this.saveFilterDropdown.bind(this);
-    this.loadFilters = this.loadFilters.bind(this);
-    this.loadSavedFilter = this.loadSavedFilter.bind(this);
   }
 
   componentDidMount() {
@@ -80,14 +60,14 @@ class SearchBoxWithFilter extends React.Component {
     }
   }
 
-  saveFilter() {
+  saveFilter = () => {
     this.props.createFilter({
       filter_name: this.state.filterName,
       filter_attributes: this.state.filters
     })
   }
 
-  loadFilters() {
+  loadFilters = () => {
     if (!this.state.loadFiltersDropdown) {
       // load filters hasn't been clicked
       this.props.loadFilters()
@@ -96,12 +76,12 @@ class SearchBoxWithFilter extends React.Component {
     this.setState({loadFiltersDropdown: !this.state.loadFiltersDropdown});
   }
 
-  loadSavedFilter(filterId) {
+  loadSavedFilter = (filterId) => {
     const savedFilter = this.props.filters.byId[filterId];
     this.setState({filters: savedFilter.filter, filterName: savedFilter.name});
   }
 
-  getPossibleFilters() {
+  getPossibleFilters = () => {
     var attribute_dict = {};
     var item_list = this.props.item_list;
 
@@ -143,104 +123,25 @@ class SearchBoxWithFilter extends React.Component {
     return attribute_dict
   }
 
-  keyNameChange(name, value) {
-    var keyNameValues = this.state.possibleFilters[value];
-
-    if (keyNameValues !== undefined) {
-      // resets keyName and keyNameValues to default to avoid controlled/uncontrolled checkbox error
-      // maps over new keyNameValues,
-
-      this.setState(
-        {
-          [name]: 'select',
-          keyNameValues: {},
-          filterType: 'of',
-          GtLtThreshold: 0,
-          dropdownActive: false
-        }, () => {
-        let attributeListIsFullyFloatParsable = true;
-        keyNameValues.map(i => {
-          if (isNaN(parseFloat(i))) {attributeListIsFullyFloatParsable = false}
-          this.setState(prevState => ({
-            [name]: value,
-            keyNameValues: {
-              ...prevState.keyNameValues,
-              [i]: false
-            }
-          }))
-        })
-        this.setState({attributeListIsFullyFloatParsable})
-      });
-    }
+  onFiltersChanged = (filters) => {
+    this.setState({
+      filters
+    })
   }
 
-  filterTypeChange(name, value) {
-    this.setState({filterType: value})
-  }
-
-  handleChange(evt) {
+  handleChange = (evt) => {
     this.setState({ [evt.target.name]: evt.target.value });
   }
 
-  addFilter() {
-    let id = this.state.filters.length + 1;
-
-    if (this.state.filterType === 'of') {
-      var newFilter = {
-        'id': id,
-        'type': 'of',
-        'keyName': this.state.keyName,
-        'allowedValues': this.get_selected_ids_array(this.state.keyNameValues)
-      }
-    } else {
-      newFilter = {
-        'id': id,
-        'type': this.state.filterType,
-        'keyName': this.state.keyName,
-        'threshold': parseFloat(this.state.GtLtThreshold)
-      }
-    }
-
-    this.setState(
-      {filters: [...this.state.filters, newFilter]},
-      () => {
-        this.setState({
-          keyName: 'select', value: 'select', keyNameValues: {}, filterType: 'of', GtLtThreshold: 0, dropdownActive: false
-        })
-        this.props.onFiltersChanged(this.state.filters)
-      }
-    )
-  }
-
-  removeFilter(evt) {
-    let newFilters = [...this.state.filters].filter(filter => (filter.id !== parseInt(evt.target.name)));
-    this.setState({filters: newFilters}, () => this.props.onFiltersChanged(this.state.filters))
-  }
-
-  toggleSelected(key) {
-    const value = !this.state.keyNameValues[key];
-
-    this.setState(prevState => ({
-      keyNameValues: {
-        ...prevState.keyNameValues,
-        [key]: value
-      },
-    }))
-  }
-
-  toggleFilter() {
+  toggleFilter = () => {
     this.setState({filterActive: !this.state.filterActive})
   }
 
-  dropdownActive() {
-    this.setState({dropdownActive: !this.state.dropdownActive})
-  }
-
-  saveFilterDropdown() {
+  saveFilterDropdown = () => {
     this.setState({saveFilterDropdown: !this.state.saveFilterDropdown})
   }
 
-  applyFilter(item_list, filter) {
+  applyFilter = (item_list, filter) => {
     return item_list.reduce((filtered, item) => {
 
       let added = false;
@@ -297,28 +198,12 @@ class SearchBoxWithFilter extends React.Component {
     }, []);
   }
 
-  get_selected_ids_array(selected) {
-    Object.filter = (obj, predicate) => Object.keys(obj)
-      .filter( key => predicate(obj[key]) ).reduce( (res, key) => (res[key] = obj[key], res), {} );
-
-    return Object.keys(Object.filter(selected, selected => selected === true));
-  }
-
   render() {
     const {
-      phrase, filters, keyName, value,
-      filterType, filterActive, possibleFilters,
-      keyNameValues, attributeListIsFullyFloatParsable, saveFilterDropdown
+      phrase, filters, filterActive, saveFilterDropdown
     } = this.state;
 
     var item_list = this.props.item_list;
-
-    let rowValues = Object.values(keyNameValues);
-    let numberSelected = rowValues.filter(Boolean).length;
-    let isSelected = numberSelected > 0 || filterType !== 'of';
-
-    // get possible filter keys, remove PROFILE_PICTURE
-    const keys = (possibleFilters !== undefined && possibleFilters !== null ? Object.keys(possibleFilters).filter(key => (key !== 'profile_picture')) : []);
 
     // Phrase Search
     if (phrase !== '') {
@@ -329,120 +214,6 @@ class SearchBoxWithFilter extends React.Component {
       this.state.filters.map(filter => {
         item_list = this.applyFilter(item_list, filter)
       });
-    }
-
-    if (isSelected) {
-      var addFilterBtn =
-        <div>
-          <StyledButton style={{fontWeight: '400', margin: '0em 1em', lineHeight: '25px', height: '25px'}} onClick={this.addFilter}>Add</StyledButton>
-        </div>
-    }
-
-
-    if (keyName !== 'select')
-    {
-      if (!attributeListIsFullyFloatParsable) {
-        var filter_type_picker = <FilterText style={{padding: '0 10px'}}>is one of</FilterText>
-      } else {
-        filter_type_picker = (
-          <StyledSelectKey name="keyName" value={filterType} onChange={(evt) => this.filterTypeChange(evt.target.name, evt.target.value)}>
-            <option name='value' value={"<"}>is less than</option>
-            <option name='value' value={"of"}>is one of</option>
-            <option name='value' value={">"}>is greater than</option>
-          </StyledSelectKey>
-        )
-      }
-    }
-
-    if (keyName !== 'select' && filterType === 'of') {
-
-      var valuePicker =
-        <div style={{display: 'flex', flexDirection: 'row', alignItems: 'center'}}>
-          <div style={{width: '200px'}}>
-            <div style={{width: 'inherit', position: 'relative'}} onClick={this.dropdownActive}>
-              <StyledSelectKey style={{width: 'inherit'}} name="value" value={value} onClick={this.dropdownActive} onChange={this.handleChange}>
-                <option name="value" value="select" disabled>select value</option>
-                {/*{typeof(custom_attribute_dict[keyName]) !== 'undefined' ? custom_attribute_dict[keyName].map((key, index) => {return (<option name='value' value={key} key={index}>{key}</option>)}) : null}*/}
-              </StyledSelectKey>
-              <div style={{position: 'absolute', top: 0, right: 0, bottom: 0, left: 0}}/>
-            </div>
-            <Checkboxes
-              style={{display: (this.state.dropdownActive ? 'block' : 'none')}}
-              onMouseLeave={() => this.setState({dropdownActive: !this.state.dropdownActive})}>
-              {typeof(possibleFilters[keyName]) !== 'undefined' ? possibleFilters[keyName].map((key, index) => (
-                  <CheckboxLabel key={index}>
-                    <input type="checkbox" value={key} checked={keyNameValues[key]} onChange={() => this.toggleSelected(key)}/>
-                    {replaceUnderscores(key)}
-                  </CheckboxLabel>
-                )) : null}
-            </Checkboxes>
-            <CloseWrapper
-            onClick={() => this.setState({dropdownActive: !this.state.dropdownActive})}
-            style={{display: (this.state.dropdownActive ? '' : 'none')}}/>
-          </div>
-
-          {addFilterBtn}
-        </div>
-    } else if (keyName !== 'select') {
-      valuePicker =
-        <div style={{display: 'flex', flexDirection: 'row', alignItems: 'center'}}>
-          <ThresholdInput type="number" name="GtLtThreshold" value={this.state.GtLtThreshold} onChange={this.handleChange}/>
-          {addFilterBtn}
-        </div>
-    }
-
-    if (filterActive) {
-      var newFilterSection =
-        <div style={{margin: '1em', display: 'flex', flexDirection: 'row', alignItems: 'center', flexFlow: 'row wrap'}}>
-
-          <div style={{display: 'flex', flexDirection: 'row', alignItems: 'center'}}>
-            <FilterText style={{padding: '0 10px 0 0'}}>Filter:</FilterText>
-            <StyledSelectKey name="keyName" value={keyName} onChange={(evt) => this.keyNameChange(evt.target.name, evt.target.value)}>
-              <option name="key" value="select" disabled>select attribute</option>
-              {typeof(keys) !== 'undefined' ? keys.map((key, index) =>
-                 <option name='value' value={key} key={index}>{replaceUnderscores(key)}</option>
-              ) : null}
-            </StyledSelectKey>
-          </div>
-          {filter_type_picker}
-          {valuePicker}
-
-        </div>
-    } else {
-      newFilterSection = null;
-    }
-
-    if (filters) {
-      var addedFilters =
-        <div style={{display: 'flex', flexDirection: 'row', alignItems: 'center', margin: '0 1em', flexFlow: 'row wrap'}}>
-          {filters.map((filter, index) => {
-            if (filter.type === 'of') {
-              return (
-                <FilterBubble key={index}>
-                  <FilterText style={{color: '#FFF'}}>
-                    {filter.keyName}: {filter.allowedValues.map((value, index) => {
-                      if (filter.allowedValues.length === index+1) {
-                        return value
-                      } else {
-                        return value + ' or '
-                      }
-                    })}
-                  </FilterText>
-                  <SVG name={filter.id} onClick={this.removeFilter} src="/static/media/close.svg"/>
-                </FilterBubble>
-              )
-            } else {
-              return (
-                <FilterBubble key={index}>
-                  <FilterText style={{color: '#FFF'}}>
-                    {filter.keyName} {filter.type} {filter.threshold}
-                  </FilterText>
-                  <SVG name={filter.id} onClick={this.removeFilter} src="/static/media/close.svg"/>
-                </FilterBubble>
-              )
-            }
-          })}
-        </div>
     }
 
     if (this.props.filters.loadStatus.isRequesting) {
@@ -523,8 +294,7 @@ class SearchBoxWithFilter extends React.Component {
         }
 
         {savedFilters}
-        {addedFilters}
-        {newFilterSection}
+        {filterActive && <Filter possibleFilters={this.state.possibleFilters} onFiltersChanged={this.onFiltersChanged}/>}
 
         <div>{React.cloneElement(this.props.children, { item_list: item_list })}</div>
       </div>
@@ -575,49 +345,15 @@ const FilterWrapper = styled.div`
   }
 `;
 
-const StyledSelectKey = styled(StyledSelect)`
-  box-shadow: 0 0 0 1px rgba(44,45,48,.15);
-  font: 400 12px system-ui;
-  color: #777;
-  padding: 0 0 0 10px;
-  margin: 5px;
-  line-height: 25px;
-  height: 25px;
-  
-`;
-
 const ThresholdInput = styled(Input)`
   font: 400 12px system-ui;
   border-radius: 5px;
   height: 12px;
 `;
 
-const FilterBubble = styled.div`
-  display: flex;
-  align-items: center;
-  width: fit-content;
-  margin: 10px 10px 0 0;
-  font: 400 12px system-ui;
-  color: #fff;
-  padding: 5px 0;
-  background-color: #607D8B;
-  border-radius: 10px;
-`;
-
 const SVG = styled.img`
   width: 12px;
   padding: 0px 10px;
-`;
-
-const Checkboxes = styled.div`
-  display: block;
-  position: absolute;
-  z-index: 55;
-  background-color: rgb(255, 255, 255);
-  width: inherit;
-  border-radius: 2px;
-  margin: 5px;
-  box-shadow: 0 0 0 1px rgba(44,45,48,.15), 0 5px 10px rgba(44,45,48,.12);
 `;
 
 const CheckboxLabel = styled.label`
